requirements.txt
flask==3.0.2
psycopg2-binary==2.9.9
pandas==2.2.1
google-generativeai==0.3.2
python-dotenv==1.0.1
plotly==5.19.0
flask-cors==4.0.0
SQLAlchemy 